刪除不需要文件

+++ src/main/java/com/power/assistant/core/service/OrgService.java
@@ -1,19 +1,28 @@
 package com.power.assistant.core.service;
 
+import com.power.assistant.mapper.MemberMapper;
 import com.power.assistant.mapper.OrgMapper;
+import com.power.assistant.mapper.UserMapper;
+import com.power.assistant.model.MemberVo;
 import com.power.assistant.model.Org;
 import com.power.assistant.model.OrgVo;
+import com.power.assistant.model.UserVo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Service
 public class OrgService {
 
     @Autowired
     private OrgMapper orgMapper;
+    @Autowired
+    private MemberMapper memberMapper;
+    @Autowired
+    private UserMapper userMapper;
 
     public List<OrgVo> selectOrgs() {
         return orgMapper.selectOrgDetailInfo(new HashMap<>());
@@ -21,6 +30,7 @@ public class OrgService {
 
     public int saveOrUpdate(Org org) {
         if (org.getId() == null) {
+            org.setStatus(1L);
             return orgMapper.insertSelective(org);
         } else {
             return orgMapper.updateById(org);
@@ -36,6 +46,19 @@ public class OrgService {
     }
 
     public int delete(Long orgId) {
+
+        //查询是否有党员或者操作员
+        HashMap<String, Object> queryMember = new HashMap<>();
+        queryMember.put("orgId",orgId);
+        List<MemberVo> memberVos = memberMapper.selectMemberInfo(queryMember);
+        if(memberVos != null && memberVos.size() > 0) {
+            throw new RuntimeException("该机构下有党员,不能被删除");
+        }
+        List<UserVo> userVos = userMapper.selectUserInfo(queryMember);
+        if(userVos != null && userVos.size() > 0) {
+            throw new RuntimeException("该机构下有操作员,不能被删除");
+        }
+
         return orgMapper.deleteById(orgId);
     }
 }
